+++ tests/Predis/Command/Utils/CommandUtilityTest.php
@@ -66,6 +66,23 @@ class CommandUtilityTest extends PredisTestCase
         CommandUtility::xxh3Hash('value');
     }
 
+    /**
+     * @return void
+     */
+    public function testDictionaryToArray(): void
+    {
+        $dict = [
+            'key1' => 'value1',
+            'key2' => 'value2',
+            'key3' => 'value3',
+        ];
+
+        $this->assertSame(
+            ['key1', 'value1', 'key2', 'value2', 'key3', 'value3'],
+            CommandUtility::dictionaryToArray($dict)
+        );
+    }
+
     public function arrayProvider(): array
     {
         return [